Added a starter code to test mircoprocessor found a link for code that seems useful for rasberry pi
Put together some code that seems like it will work for aurdino bluetooth transmitting. It reads an input from it's ADC using analog read
then outputs it using bluetooth print. This device has a default password of 1234 or 0000 and this can be changed by essentially changing its
fuses to command mode then using special commands to change it, which we will want to do later but for now it is fine as it is.

For the pi, I looked into it a little bit and it seems like I will have to use some linux commands to pair with the device before the two can
communicate. I am going to look into puting this into a config file or a bash script to auto run with my gui but for now I will just run it
and write down all the commands that I need to use. I found an article that seems like it will be helpful for the pairing process so I put that into the
starter_code.py file