Me and Will got Pi to recieve data from arduino using bluetooth connection
We had to play with each of the codes a little bit but we were able to get the rasberry pi to
connect to the arduino and send a premade message from arduino to pi. Right now it is a little
finicky and the arduino needs to be turned off and back on if the connection is lost but overall
we can get it to work consistently. To run it the devices will pair automatically at the moment
so all we need to do is turn on both devices, then do sudo rfcomm connect hci0 xxxxxxxx, where x's
are the device number of the bluetooth device, which can be found using hcitool scan. Will knows
the wiring that needs to be done for connecting the bluetooth module to the arduino.