Fixed a typo in one of the placeholder bash scripts
Fixed a typo in one of the placeholder bash scripts. To make the gui working
for the pop project I am going to need to change the go to rather than showing
and hiding the graph it shows the graph only once the data has been collected.
I also will need a way to make the script to repeatedly try to start the bluetooth
connect command over and over be independent of the gui thread, either by
some bash method or a python multithreading method.