Worked on the gui a little bit more. Can start and stop bash process from GUI
Worked on the tkinter gui a little bit more. Figured out that I will need to use python subprocesses using Popen.
Eventually got a nifty solution to be able to kill a thread once it is created by having a class for the button that includes the button as well as
the variable where the subprocess will be stored, then I can return that on the button push to allow the gui code to work like normal when the button is pushed,
but still be able to kill the data stream when transmission is done. Might still be leaving an orphan becuase I'm not 100% sure if I am killing the process or just pausing it.
I might need to wait for it after I kill it.

In the future we will be able to have the bash script take standard output from the other python code and store it in a file via tee, then we can acess that file once we kill
the subprocess to get the data collected during transmission.