Wrote sick python code for the xbee module
I love not having to use the garbage bluetooth module anymore because
I can just hook the XBee module up to my laptop and code on my computer.
The code reads 20 packets of Data then graphs them.
There's even code that I think might make a spectrum of the data.

I figured out a method for the samples overlapping packets issue that I
will explain in the readme

The issue that I have seen so faw is when I was working on it outside a few packets got lost (failed checksum)