Added logic for different data types, fixed graphs
Wrote logic that works for different data types assuming the simplest
solution in which after each data type is sent, it sends a special char
indicating what the data is. I'm pretty sure that the channel will be
able to accomodate this since the other two types are slow hertz

Fixed the subplots to have correct headers for whats stored on them and
fixed the layout so that they wouldn't overlap. In matplotlib unlike
matlab plot starts out with hold on enabled as a deault which is super
cool