Got multi channel transmission to work!!!
For python code all I changed was adding the correct sampling frequencies to the xaxis on the graphs
and I added the other special chars to one of the if statements

On the arduino side I wrote code where I check for counter matches in
the ISR that indicate its time to sample the other channels. I then
have the status var change to a different var based on the match.
Then in the print section I have a switch case for the different status
values.

Also I learned that sei() and cli() actaully do something. When
interrupts aren't disabled the vars get printed with \ns, but with
interrupts enabled they print all in one line with no sampling errors.
Not really sure why that happens but pretty cool :-).